src/index.ts
/**
 * Provides reflection capabilities for dynamic types.
 */
export class DynamicTypeReflection {

    private static typeRegistry: Set<Record<string, string>> = new Set();

    /**
     * Registers a type to the runtime system.
     * @param type The type to be registered.
     */
    public static registerType(type: Record<string, string>): void {
        this.typeRegistry.add(type);
    }

    /**
     * Checks if a type has a valid definition in the runtime system.
     * @param type The type to check.
     * @returns Boolean indicating validity.
     */
    public static hasValidDefinition(type: Record<string, string>): boolean {
        return this.typeRegistry.has(type);
    }

    /**
     * Returns the properties of a given type.
     * @param type The type to inspect.
     * @returns Array of property names.
     */
    public static getProperties(type: Record<string, string>): string[] {
        return Object.keys(type);
    }
}

/**
 * Offers utilities for runtime type validation.
 */
export class DynamicTypeValidator {

    /**
     * Validates if an instance matches the type's structure.
     * @param instance The object instance.
     * @param type The expected type.
     * @returns Boolean indicating if instance matches type.
     */
    public static validateType(instance: any, type: any): boolean {
        // Check if properties in type match instance
        for (let key in type) {
            if (!(key in instance) || typeof instance[key] !== type[key]) {
                return false;
            }
        }

        // Check if instance has properties not in type
        for (let key in instance) {
            if (!(key in type)) {
                return false;
            }
        }

        return true;
    }
}

/**
 * Contains utilities for code generation based on templates.
 */
export class CodeGenerator {

    /**
     * Generates code based on a template with type constraints.
     * @param template The code template.
     * @param typeConstraints Constraints for the code.
     * @returns Generated code as a string.
     */
    public static generateCode(template: string, typeConstraints: Record<string, string>): string {
        let code = template;
        for (let key in typeConstraints) {
            const placeholder = `{{${key}}}`;
            code = code.replace(new RegExp(placeholder, 'g'), typeConstraints[key]);
        }
        return code;
    }
}

/**
 * Provides a type-safe wrapper for third-party libraries.
 */
export class ThirdPartyTypeWrapper {

    /**
     * Wraps a library with a type-safe proxy.
     * @param lib The third-party library.
     * @returns Type-safe proxy around the library.
     */
    /**
      * Wraps a library with a type-safe proxy.
      * @param lib The third-party library.
      * @returns Type-safe proxy around the library.
      */
    public static wrapWithProxy(lib: any): any {
        return new Proxy(lib, {
            get(target, prop, receiver) {
                if (prop in target) {
                    return Reflect.get(target, prop, receiver);
                } else {
                    throw new Error(`Access to property ${String(prop)} is not allowed for type safety.`);
                }
            },
            set(target, prop, value) {
                if (prop in target) {
                    return Reflect.set(target, prop, value);
                } else {
                    throw new Error(`Setting new property ${String(prop)} is not allowed for type safety.`);
                }
            }
        });
    }
}

/**
 * Main library class to create dynamic types.
 */
export class DynamicTypeSafe {

    /**
     * Creates a dynamic type based on the provided schema.
     * @param schema The structure definition.
     * @returns A constructor function for the dynamic type.
     */
    public static createDynamicType(schema: Record<string, string>): new () => any {
        // Validate the provided schema
        const validTypes = ['string', 'number', 'boolean', 'object', 'function', 'symbol', 'undefined'];
        for (const key in schema) {
            if (!validTypes.includes(schema[key])) {
                throw new Error(`Invalid type descriptor for property ${key}: ${schema[key]}`);
            }
        }

        class DynamicClass {
            constructor() {
                for (let key in schema) {
                    (this as any)[key] = null;
                }
            }
        }

        return DynamicClass;
    }
}
